+++ Lab1/src/point/Point.java
@@ -4,20 +4,20 @@ import java.util.Objects;
 
 public final class Point {
 
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     // no need for copies as they are already immutable
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() { return y; }
+    public double getY() { return y; }
 
     @Override
     public String toString() {
